ios: update pns handling for bonus related pns, add #2052

+++ app-ios-native/qingshow-ios-native/libs/CKCalendarView/CKCalendarView.m
@@ -452,7 +452,7 @@
 
 - (int)dayOfWeekForDate:(NSDate *)date {
     NSDateComponents *comps = [self.calendar components:NSWeekdayCalendarUnit fromDate:date];
-    return comps.weekday;
+    return (int)comps.weekday;
 }
 
 - (BOOL)dateIsToday:(NSDate *)date {
